refactor(🔧): Restructure directories
