--- src/components/Navbar/tests/index.test.tsx
@@ -1,3 +1,0 @@
-describe("Navbar", () => {
-  test("should contain menu button, leaderboard button and logo button", () => {});
-});
