+++ src/containers/Game/PlayMatLayout.tsx
@@ -4,7 +4,7 @@ interface PropsWithChildren {
     children: ReactNode;
 }
 
-export default function Layout(props: PropsWithChildren): ReactElement {
+export default function PlayMatLayout(props: PropsWithChildren): ReactElement {
     const { children } = props;
     return (
         <div className="relative flex-1 flex flex-row-reverse tablet:flex-col w-full pr-20 tablet:pr-0">
