+++ src/containers/Welcome/Infographic.tsx
