+++ src/pages/WelcomePage.tsx
@@ -1,5 +1,6 @@
 import React from "react";
 import {Welcome} from "../containers/Welcome";
+
 export default function WelcomePage(){
     return (<Welcome/>)
 }
