+++ src/providers/CreateGameProvider/useCreateGameContext.ts
@@ -5,7 +5,7 @@ export default function useCreateGameContext() {
   const context = useContext(CreateGameContext);
   if (!context) {
     throw new Error(
-      "useCreateOrJoinGameContext must be used within a CreateorJoinGameProvider"
+      "useCreateGameContext must be used within a CreateGameProvider"
     );
   }
   return context;
